fix: harden QRCode package release surface
