+++ package.json
@@ -9,20 +9,21 @@
     "setup": "bun scripts/setup.js",
     "publish-package": "bun scripts/publish.js",
     "publish-package:dry-run": "bun scripts/publish.js --dry-run --skip-checks",
-    "build": "turbo run build",
+    "build": "bun run --cwd packages/react-native-nitro-qrcode build",
     "dev": "bun run example:start",
-    "check": "bun run lint:check && bun run typecheck && bun run test:coverage && bun run test:cpp && bun run build && bun run example:check",
+    "check": "bun run lint:check && bun run typecheck && bun run test:types && bun run test:coverage && bun run test:cpp && bun run build && bun run example:check",
     "check:ci": "bun run check && bun run test:cpp:sanitize",
-    "release:preflight": "bun run check:ci && bun run audit:package && bun run publish-package:dry-run",
-    "lint": "turbo run lint",
-    "lint:check": "turbo run lint:check",
-    "clean": "turbo run clean",
-    "codegen": "turbo run codegen",
-    "typecheck": "turbo run typecheck",
-    "test": "turbo run test",
-    "test:unit": "turbo run test",
-    "test:coverage": "turbo run test:coverage",
-    "test:cpp": "turbo run test:cpp --filter=react-native-nitro-qrcode",
+    "release:preflight": "bun run check:ci && bun run benchmark:cpp && bun run audit:package && bun run publish-package:dry-run",
+    "lint": "bun run --cwd packages/react-native-nitro-qrcode lint",
+    "lint:check": "bun run --cwd packages/react-native-nitro-qrcode lint:check",
+    "clean": "bun run --cwd packages/react-native-nitro-qrcode clean",
+    "codegen": "bun run --cwd packages/react-native-nitro-qrcode codegen",
+    "typecheck": "bun run --cwd packages/react-native-nitro-qrcode typecheck",
+    "test:types": "bun run --cwd packages/react-native-nitro-qrcode test:types",
+    "test": "bun run --cwd packages/react-native-nitro-qrcode test",
+    "test:unit": "bun run --cwd packages/react-native-nitro-qrcode test",
+    "test:coverage": "bun run --cwd packages/react-native-nitro-qrcode test:coverage",
+    "test:cpp": "bun run --cwd packages/react-native-nitro-qrcode test:cpp",
     "test:cpp:sanitize": "bun scripts/test-cpp-sanitize.js",
     "benchmark:cpp": "bun scripts/benchmark-cpp.js",
     "audit:package": "bun scripts/audit-package-contents.js",
@@ -63,7 +64,6 @@
     "react-test-renderer": "19.2.3",
     "rimraf": "^6.1.3",
     "ts-jest": "^29.4.11",
-    "turbo": "^2.9.14",
     "typescript": "^6.0.3"
   },
   "overrides": {
@@ -71,7 +71,7 @@
     "react-dom": "19.2.3",
     "react-native": "0.85.3",
     "react-native-nitro-modules": "0.35.7",
-    "expo-constants": "~56.0.17",
+    "expo-constants": "~56.0.18",
     "@types/react": "~19.2.15"
   },
   "engines": {
