+++ packages/react-native-nitro-qrcode/package.json
@@ -58,12 +58,13 @@
     "lint": "bunx eslint src --max-warnings=0 --ignore-pattern 'src/**/*.nitro.ts'",
     "lint:check": "bunx eslint src --max-warnings=0 --ignore-pattern 'src/**/*.nitro.ts'",
     "typecheck": "tsc --noEmit",
+    "test:types": "tsc --noEmit -p tsconfig.types.json",
     "test": "jest",
     "test:coverage": "jest --coverage",
     "test:cpp": "bun ../../scripts/test-cpp.js",
     "test:cpp:sanitize": "bun ../../scripts/test-cpp-sanitize.js",
     "benchmark:cpp": "bun ../../scripts/benchmark-cpp.js",
-    "verify": "bun run lint:check && bun run typecheck && bun run test:coverage && bun run test:cpp && bun run test:cpp:sanitize && bun run build",
+    "verify": "bun run lint:check && bun run typecheck && bun run test:types && bun run test:coverage && bun run test:cpp && bun run test:cpp:sanitize && bun run build",
     "prepublishOnly": "bun run clean && bun run codegen && bun run verify",
     "prepack": "node -e \"const fs=require('fs'); fs.copyFileSync('../../README.md','./README.md'); try{fs.copyFileSync('../../LICENSE','./LICENSE')}catch(e){}\"",
     "postpack": "node -e \"const fs=require('fs'); ['./README.md','./LICENSE'].forEach(f=>fs.existsSync(f)&&fs.unlinkSync(f))\"",
