+++ packages/react-native-nitro-qrcode/tsconfig.json
@@ -3,9 +3,27 @@
   "compilerOptions": {
     "outDir": "./lib/typescript",
     "rootDir": "./src",
-    "lib": ["ES2020", "DOM"],
-    "types": ["react-native"]
+    "lib": [
+      "ES2020",
+      "DOM"
+    ],
+    "types": [
+      "react-native"
+    ],
+    "noUncheckedIndexedAccess": true,
+    "exactOptionalPropertyTypes": false,
+    "noImplicitOverride": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true,
+    "useUnknownInCatchVariables": true
   },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "lib", "nitrogen", "src/__tests__"]
+  "include": [
+    "src/**/*"
+  ],
+  "exclude": [
+    "node_modules",
+    "lib",
+    "nitrogen",
+    "src/__tests__"
+  ]
 }
