+++ packages/react-native-nitro-qrcode/tsconfig.types.json
@@ -1,0 +1,10 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "exclude": [],
+  "include": [
+    "src/**/*.test-d.ts"
+  ]
+}
