+++ scripts/audit-package-contents.js
@@ -41,14 +41,24 @@ const requiredFiles = [
   "cpp/core/QRCodeGenerator.cpp",
   "cpp/qrcodegen/qrcodegen.cpp",
   "lib/commonjs/index.js",
+  "lib/commonjs/qrcode-component.js",
+  "lib/commonjs/shared.js",
   "lib/module/index.js",
+  "lib/module/qrcode-component.js",
+  "lib/module/shared.js",
   "lib/typescript/commonjs/index.d.ts",
+  "lib/typescript/commonjs/qrcode-component.d.ts",
+  "lib/typescript/commonjs/shared.d.ts",
   "lib/typescript/module/index.d.ts",
+  "lib/typescript/module/qrcode-component.d.ts",
+  "lib/typescript/module/shared.d.ts",
   "nitrogen/generated/android/NitroQRCode+autolinking.cmake",
   "nitrogen/generated/ios/NitroQRCode+autolinking.rb",
   "nitrogen/generated/shared/c++/HybridQRCodeSpec.hpp",
   "src/index.ts",
   "src/index.web.ts",
+  "src/qrcode-component.ts",
+  "src/shared.ts",
 ];
 
 const forbiddenPatterns = [
