+++ scripts/test-cpp-sanitize.js
@@ -1,4 +1,4 @@
-const { execSync } = require("child_process");
+const { execFileSync, execSync } = require("child_process");
 const fs = require("fs");
 const path = require("path");
 
@@ -12,6 +12,26 @@ const cppDir = path.join(packageDir, "cpp");
 const buildDir = path.join(cppDir, "build-sanitize");
 const outputFile = path.join(buildDir, "qrcode_generator_test_sanitize");
 
+function resolveTool(name) {
+  try {
+    return execSync(`command -v ${name}`, { encoding: "utf8" }).trim();
+  } catch {
+    if (process.platform !== "darwin") {
+      throw new Error(`${name} was not found on PATH.`);
+    }
+
+    return execFileSync("xcrun", ["--find", name], {
+      encoding: "utf8",
+    }).trim();
+  }
+}
+
+function runCommand(command, args) {
+  execFileSync(command, args, {
+    stdio: "inherit",
+  });
+}
+
 fs.rmSync(buildDir, { recursive: true, force: true });
 fs.mkdirSync(buildDir, { recursive: true });
 
@@ -23,8 +43,7 @@ const sources = [
   path.join(cppDir, "qrcodegen", "qrcodegen.cpp"),
 ];
 
-const compileCmd = [
-  "clang++",
+const compileArgs = [
   "-std=c++20",
   "-Wall",
   "-Wextra",
@@ -41,10 +60,10 @@ const compileCmd = [
   outputFile,
   "-lz",
   process.platform === "darwin" ? "-stdlib=libc++" : "-lpthread",
-].join(" ");
+];
 
 console.log("Compiling C++ QRCode tests with ASan/UBSan...");
-execSync(compileCmd, { stdio: "inherit" });
+runCommand(resolveTool("clang++"), compileArgs);
 
 console.log("Running C++ QRCode sanitizer tests...");
-execSync(outputFile, { stdio: "inherit" });
+runCommand(outputFile, []);
