--- turbo.json
@@ -1,65 +1,0 @@
-{
-  "$schema": "https://turbo.build/schema.json",
-  "globalDependencies": [
-    "package.json",
-    "bun.lock",
-    "bunfig.toml",
-    "eslint.config.js",
-    "tsconfig.json",
-    "packages/*/package.json",
-    "apps/*/package.json"
-  ],
-  "globalPassThroughEnv": ["ANDROID_HOME", "ANDROID_SDK_ROOT", "JAVA_HOME"],
-  "tasks": {
-    "build": {
-      "dependsOn": ["^build"],
-      "outputs": ["lib/**", "dist/**"]
-    },
-    "lint": {
-      "dependsOn": ["^build"]
-    },
-    "lint:check": {
-      "dependsOn": ["^build"]
-    },
-    "clean": {
-      "cache": false
-    },
-    "codegen": {
-      "cache": false,
-      "outputs": ["nitrogen/generated/**"]
-    },
-    "typecheck": {
-      "dependsOn": ["^build"]
-    },
-    "test": {
-      "dependsOn": ["^build"],
-      "outputs": ["coverage/**"]
-    },
-    "test:coverage": {
-      "dependsOn": ["^build"],
-      "outputs": ["coverage/**"]
-    },
-    "prebuild": {
-      "cache": false
-    },
-    "prebuild:clean": {
-      "cache": false
-    },
-    "ios": {
-      "cache": false
-    },
-    "android": {
-      "cache": false
-    },
-    "start": {
-      "cache": false,
-      "persistent": true
-    },
-    "test:cpp": {
-      "cache": false
-    },
-    "test:cpp:sanitize": {
-      "cache": false
-    }
-  }
-}
